fix(low): 批量修复 19 个 LOW 问题
后端：
- 移除 query_service 死代码 _sticky_condition_base
- mobile/articles.py print→logger
- articles.py 添加 db.refresh 确保数据最新
- revision compar 标签使用版本号
- 集成 AuditLogService 到登录/2FA/注册
- SECRET_KEY 统一使用 secrets.token_urlsafe
- category parent_id 添加 FK SET NULL
- revision_number BigInteger→Integer
- mobile 视图计数添加 IP 冷却
- V3 admin 添加 is_staff 检查（与 V2 一致）
- 会话 ID 改用 secrets.token_hex
- _remove_session 改为实际移除列表元素
- XSS/SQLi 中间件排除路径 /api/v1/→/api/v2/
- 2FA 设置添加多进程警告日志
- shortcode note_type 白名单验证
- field_permission ThreadPoolExecutor 添加 TODO

前端：
- PermissionGuard useCapability fail-open→fail-closed
- 移除 production console.log
- ArticleEditorPage 添加注释说明预览限制

+++ shared/models/article/article_revision.py
@@ -29,7 +29,7 @@ class ArticleRevision(Base):
     article_id = Column(BigInteger, ForeignKey('articles.id'), doc='关联的文章ID')
 
 
-    revision_number = Column(BigInteger, doc='修订版本号')
+    revision_number = Column(Integer, doc='修订版本号')
 
 
     title = Column(String(255), nullable=True, doc='修订时的标题')
